facebook table update

+++ database/migrations/2019_05_21_091223_create_social_facebook_accounts_table.php
@@ -15,7 +15,7 @@ class CreateSocialFacebookAccountsTable extends Migration
     {
         Schema::create('social_facebook_accounts', function (Blueprint $table) {
             $table->integer('user_id');
-            $table->integer('provider_user_id');
+            $table->bigint('provider_user_id');
             $table->string('provider');
             $table->timestamps();
         });
